move OOP object to forward_object header
